Initial Improvements
1. The script was transferred to the Task1 directory.
2. The "extracting_data" function was restructured, which resulted in a significant reduction in the number of lines of code and increased readability.
3. All xpaths were relocated to a separate config file.
4. Additional comments were added to improve understanding.
