+++ Task_1/XPath.config
@@ -1,0 +1,10 @@
+{
+    "artist_name": "//*[@id=\"main_center_0_lblLotPrimaryTitle\"]/text()",
+    "painting_name": "//*[@id=\"main_center_0_lblLotSecondaryTitle\"]/.//text()",
+    "price_gbp": "//*[@id=\"main_center_0_lblPriceRealizedPrimary\"]/text()",
+    "price_usd": "//*[@id=\"main_center_0_lblPriceRealizedSecondary\"]/text()",
+    "price_gbp_est": "//*[@id=\"main_center_0_lblPriceEstimatedPrimary\"]/text()",
+    "price_usd_est": "//*[@id=\"main_center_0_lblPriceEstimatedSecondary\"]/text()",
+    "image_link": "//*[@id=\"imgLotImage\"]/@src",
+    "sale_date": "//*[@id=\"main_center_0_lblSaleDate\"]/text()"
+}
